// ClientApp/FxWin.ClientApp/ClientApp/app/core/models/enums.ts
export enum enumHedgeType {
    Buy = 1,
    Sale,
    Swap,
}

export enum enumPurchaseType {
    Achat = 1,
    Vente,
}

export enum enumWorkFlowStatus {
    BROUILLON = 1,
    EN_COURS = 2,
    A_VERIFIER = 3,
    VALIDE = 4,
    ANNULE = 5
}


export enum ViewOpenMode {
    CreationMode = 1,
    EditionMode = 2
}


export enum enumUserRole {
    FO = 1,
    BO = 2,
    ADMIN = 4,
    R = 5,
}